move tests to ./tests
