+++ test/CMakeLists.txt
@@ -1,5 +1,7 @@
-add_executable(test_all test_all.cpp)
+add_executable(test_all test_all.cpp test_pbc_periodic.cpp)
 target_link_libraries(test_all dmk_shared ${DMK_REFERENCE_TARGET} doctest::doctest ${DMK_MPI_TARGET} ${DMK_OPENMP_TARGET} BLAS::BLAS)
+target_include_directories(test_all PRIVATE ${DMK_INCLUDES} ${PROJECT_SOURCE_DIR}/extern/nda/c++
+  ${PROJECT_BINARY_DIR}/extern/nda/deps/itertools_src/c++)
 
 add_executable(benchmark_gradient ../scripts/benchmark_gradient.cpp)
 target_link_libraries(benchmark_gradient dmk_shared nda::nda_c spdlog::spdlog_header_only ${DMK_OPENMP_TARGET} BLAS::BLAS ${DMK_MPI_TARGET})
